backend/requirements.txt
faiss-cpu==1.11.0
fastapi==0.115.12
PyMuPDF==1.25.5
pytesseract==0.3.13
python-docx==1.1.2
python-multipart==0.0.20
requests==2.32.3
sentence-transformers==4.1.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
uvicorn==0.34.2

